GraphPersistance to access internals of GraphImpl and its Scope
